+++ CHANGELOG.md
@@ -4,8 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- An event bus instance is now available by default at "window.intermix.eventBus".
+- Sequencer.js: Frame animation stall/resume when user switches browser tab.
+
 ### Changed
-- Improved README.md
+- Sequencer.js: New algorithm for frame animation that runs at least 5.5 times
+  faster (@120bpm without stall/resume). In reality, speed improvement should be even higher.
+- Sequencer.js: Some minor speed optimizations.
+- README.md improved.
+- Changelog.md improved.
 
 ## [0.3.0] - 2016-05-27
 ### Fixed